removed dependency on `splash-screen` in base module
