+++ app/build.gradle.kts
@@ -109,7 +109,6 @@ dependencies {
   implementation(Dependencies.GoogleMaterial)
   implementation("androidx.constraintlayout:constraintlayout:2.1.4")
   implementation("androidx.appcompat:appcompat:1.5.1")
-  implementation("androidx.core:core-splashscreen:1.0.0")
   implementation("androidx.core:core-ktx:1.9.0")
   implementation("androidx.fragment:fragment-ktx:1.5.4")
   implementation("androidx.navigation:navigation-fragment-ktx:2.5.3")
