+++ app/src/main/java/dev/teogor/ceres/main/MainActivity.kt
@@ -16,31 +16,17 @@
 
 package dev.teogor.ceres.main
 
-import android.animation.ValueAnimator
-import android.graphics.Color
-import android.os.Build
 import android.view.View
-import androidx.core.animation.doOnEnd
-import androidx.core.splashscreen.SplashScreen
-import androidx.core.view.ViewCompat
-import androidx.core.view.WindowInsetsCompat
 import androidx.databinding.library.baseAdapters.BR
-import androidx.interpolator.view.animation.FastOutSlowInInterpolator
-import androidx.interpolator.view.animation.LinearOutSlowInInterpolator
 import dagger.hilt.android.AndroidEntryPoint
 import dev.chrisbanes.insetter.applyInsetter
 import dev.teogor.ceres.R
 import dev.teogor.ceres.databinding.ActivityMainBinding
-import dev.teogor.ceres.extensions.dpToPx
 import dev.teogor.ceres.extensions.isVisible
-import dev.teogor.ceres.extensions.setNavigationBarTransparentCompat
 import dev.teogor.ceres.m3.app.BaseActivityM3
 
 @AndroidEntryPoint
 class MainActivity : BaseActivityM3<ActivityMainBinding, MainViewModel>() {
-
-  // To be checked by splash screen. If true then splash screen will be removed.
-  var ready = false
 
   override fun getContentView(): Int = R.layout.activity_main
 
@@ -50,22 +36,16 @@ class MainActivity : BaseActivityM3<ActivityMainBinding, MainViewModel>() {
 
   override fun getViewModelClass() = MainViewModel::class.java
 
-  override fun drawEdgeToEdge(splashScreen: SplashScreen?) {
-    super.drawEdgeToEdge(splashScreen)
+  override val splashExitAnimDuration: Long = 1000
+
+  override fun drawEdgeToEdge() {
+    super.drawEdgeToEdge()
 
     binding.bottomNavigation.applyInsetter {
       type(navigationBars = true) {
         padding()
       }
     }
-
-    val startTime = System.currentTimeMillis()
-    val elapsed = System.currentTimeMillis() - startTime
-    val keepSplashOnScreen =
-      elapsed <= SPLASH_MIN_DURATION || !ready && elapsed <= SPLASH_MAX_DURATION
-    // splashScreen?.setKeepOnScreenCondition{keepSplashOnScreen}
-
-    setSplashScreenExitAnimation(splashScreen)
   }
 
   override fun setupObservers() {
@@ -105,75 +85,7 @@ class MainActivity : BaseActivityM3<ActivityMainBinding, MainViewModel>() {
     binding.bottomNavigation.isVisible = false
   }
 
-  /**
-   * Sets custom splash screen exit animation on devices prior to Android 12.
-   *
-   * When custom animation is used, status and navigation bar color will be set
-   * to transparent and will be restored after the animation is finished.
-   */
-  private fun setSplashScreenExitAnimation(splashScreen: SplashScreen?) {
-    val setNavbarScrim = {
-      // Make sure navigation bar is on bottom before we modify it
-      ViewCompat.setOnApplyWindowInsetsListener(binding.root) { _, insets ->
-        if (insets.getInsets(WindowInsetsCompat.Type.navigationBars()).bottom > 0) {
-          val elevation = binding.bottomNavigation.elevation
-          window.setNavigationBarTransparentCompat(this@MainActivity, elevation)
-        }
-        insets
-      }
-      ViewCompat.requestApplyInsets(binding.root)
-    }
-
-    if (Build.VERSION.SDK_INT < Build.VERSION_CODES.S && splashScreen != null) {
-      val oldStatusColor = window.statusBarColor
-      val oldNavigationColor = window.navigationBarColor
-      window.statusBarColor = Color.TRANSPARENT
-      window.navigationBarColor = Color.TRANSPARENT
-
-      splashScreen.setOnExitAnimationListener { splashProvider ->
-        // For some reason the SplashScreen applies (incorrect) Y translation to the iconView
-        splashProvider.iconView.translationY = 0F
-
-        val activityAnim = ValueAnimator.ofFloat(1F, 0F).apply {
-          interpolator = LinearOutSlowInInterpolator()
-          duration = SPLASH_EXIT_ANIM_DURATION
-          addUpdateListener { va ->
-            val value = va.animatedValue as Float
-            binding.root.translationY = value * 16.dpToPx
-          }
-        }
-
-        val splashAnim = ValueAnimator.ofFloat(1F, 0F).apply {
-          interpolator = FastOutSlowInInterpolator()
-          duration = SPLASH_EXIT_ANIM_DURATION
-          addUpdateListener { va ->
-            val value = va.animatedValue as Float
-            splashProvider.view.alpha = value
-          }
-          doOnEnd {
-            splashProvider.remove()
-            window.statusBarColor = oldStatusColor
-            window.navigationBarColor = oldNavigationColor
-            setNavbarScrim()
-          }
-        }
-
-        activityAnim.start()
-        splashAnim.start()
-      }
-    } else {
-      setNavbarScrim()
-    }
-  }
-
   override fun getSnackbarAnchor(): View {
     return binding.bottomNavigation
-  }
-
-  companion object {
-    // Splash screen
-    private const val SPLASH_MIN_DURATION = 500 // ms
-    private const val SPLASH_MAX_DURATION = 5000 // ms
-    private const val SPLASH_EXIT_ANIM_DURATION = 400L // ms
   }
 }
