+++ ceres-components/src/main/kotlin/dev/teogor/ceres/components/app/BaseActivity.kt
@@ -16,8 +16,11 @@
 
 package dev.teogor.ceres.components.app
 
+import android.animation.ValueAnimator
 import android.content.Context
 import android.content.res.Configuration
+import android.graphics.Color
+import android.os.Build
 import android.os.Bundle
 import android.view.View
 import android.view.Window
@@ -25,11 +28,15 @@ import android.widget.Toast
 import androidx.annotation.IdRes
 import androidx.annotation.LayoutRes
 import androidx.appcompat.app.AppCompatActivity
+import androidx.core.animation.doOnEnd
 import androidx.core.splashscreen.SplashScreen
 import androidx.core.splashscreen.SplashScreen.Companion.installSplashScreen
+import androidx.core.view.ViewCompat
 import androidx.core.view.WindowCompat
 import androidx.databinding.DataBindingUtil
 import androidx.databinding.ViewDataBinding
+import androidx.interpolator.view.animation.FastOutSlowInInterpolator
+import androidx.interpolator.view.animation.LinearOutSlowInInterpolator
 import androidx.lifecycle.ViewModelProvider
 import androidx.navigation.NavController
 import androidx.navigation.fragment.DialogFragmentNavigator
@@ -41,8 +48,10 @@ import dev.teogor.ceres.components.events.UiEvent
 import dev.teogor.ceres.components.navigation.NavigationUI
 import dev.teogor.ceres.components.navigation.NavigationViewModel
 import dev.teogor.ceres.core.internal.WindowPreferencesManager
+import dev.teogor.ceres.extensions.dpToPx
 import dev.teogor.ceres.extensions.findNavController
 import dev.teogor.ceres.extensions.hideKeyboard
+import java.util.concurrent.TimeUnit
 
 abstract class BaseActivity<B : ViewDataBinding, VM : BaseViewModel> :
   AppCompatActivity(), BaseCommon {
@@ -55,6 +64,9 @@ abstract class BaseActivity<B : ViewDataBinding, VM : BaseViewModel> :
   lateinit var navigationViewModel: NavigationViewModel
 
   private var previousToast: Toast? = null
+
+  open val splashExitAnimDuration: Long
+    get() = TimeUnit.MILLISECONDS.toMillis(400L)
 
   // todo navController move to a module that contains firebase
   private val listener = NavController.OnDestinationChangedListener { controller, _, _ ->
@@ -99,7 +111,8 @@ abstract class BaseActivity<B : ViewDataBinding, VM : BaseViewModel> :
     binding.lifecycleOwner = this
     binding.setVariable(getBindingVariable(), viewModel)
 
-    drawEdgeToEdge(splashScreen)
+    drawEdgeToEdge()
+    setSplashScreenExitAnimation(splashScreen)
 
     getNavHostFragment()?.apply {
       navController.addOnDestinationChangedListener(listener)
@@ -124,7 +137,7 @@ abstract class BaseActivity<B : ViewDataBinding, VM : BaseViewModel> :
   @IdRes
   open fun getNavController(): Int = -1
 
-  protected open fun drawEdgeToEdge(splashScreen: SplashScreen?) {
+  protected open fun drawEdgeToEdge() {
     // Draw edge-to-edge
     WindowPreferencesManager().applyEdgeToEdgePreference(
       window
@@ -177,6 +190,63 @@ abstract class BaseActivity<B : ViewDataBinding, VM : BaseViewModel> :
     configuration.uiMode = Configuration.UI_MODE_NIGHT_UNDEFINED
     val baseContext = cxt.createConfigurationContext(configuration)
     super.attachBaseContext(baseContext)
+  }
+
+  /**
+   * Sets custom splash screen exit animation on devices prior to Android 12.
+   *
+   * When custom animation is used, status and navigation bar color will be set
+   * to transparent and will be restored after the animation is finished.
+   */
+  private fun setSplashScreenExitAnimation(splashScreen: SplashScreen?) {
+    val setNavbarScrim = {
+      // Make sure navigation bar is on bottom before we modify it
+      ViewCompat.setOnApplyWindowInsetsListener(binding.root) { _, insets ->
+        insets
+      }
+      ViewCompat.requestApplyInsets(binding.root)
+    }
+
+    if (Build.VERSION.SDK_INT < Build.VERSION_CODES.S && splashScreen != null) {
+      val oldStatusColor = window.statusBarColor
+      val oldNavigationColor = window.navigationBarColor
+      window.statusBarColor = Color.TRANSPARENT
+      window.navigationBarColor = Color.TRANSPARENT
+
+      splashScreen.setOnExitAnimationListener { splashProvider ->
+        // For some reason the SplashScreen applies (incorrect) Y translation to the iconView
+        splashProvider.iconView.translationY = 0F
+
+        val activityAnim = ValueAnimator.ofFloat(1F, 0F).apply {
+          interpolator = LinearOutSlowInInterpolator()
+          duration = splashExitAnimDuration
+          addUpdateListener { va ->
+            val value = va.animatedValue as Float
+            binding.root.translationY = value * 16.dpToPx
+          }
+        }
+
+        val splashAnim = ValueAnimator.ofFloat(1F, 0F).apply {
+          interpolator = FastOutSlowInInterpolator()
+          duration = splashExitAnimDuration
+          addUpdateListener { va ->
+            val value = va.animatedValue as Float
+            splashProvider.view.alpha = value
+          }
+          doOnEnd {
+            splashProvider.remove()
+            window.statusBarColor = oldStatusColor
+            window.navigationBarColor = oldNavigationColor
+            setNavbarScrim()
+          }
+        }
+
+        activityAnim.start()
+        splashAnim.start()
+      }
+    } else {
+      setNavbarScrim()
+    }
   }
 
   /**
